+++ lib/src/core/utils/index.dart
@@ -1,1 +1,3 @@
 export 'machine_learning_kit_helper.dart';
+export 'liveness_challenge_flow_builder.dart';
+export 'liveness_face_orientation_helper.dart';