+++ lib/src/presentation/widgets/liveness_detection_step_overlay_widget.dart
@@ -6,6 +6,7 @@ import 'package:lottie/lottie.dart';
 class LivenessDetectionStepOverlayWidget extends StatefulWidget {
   final List<LivenessDetectionStepItem> steps;
   final VoidCallback onCompleted;
+  final Duration completionDelay;
   final Widget camera;
   final CameraController? cameraController;
   final bool isFaceDetected;
@@ -18,6 +19,7 @@ class LivenessDetectionStepOverlayWidget extends StatefulWidget {
     super.key,
     required this.steps,
     required this.onCompleted,
+    this.completionDelay = const Duration(milliseconds: 500),
     required this.camera,
     required this.cameraController,
     required this.isFaceDetected,
@@ -112,9 +114,7 @@ class LivenessDetectionStepOverlayWidgetState
       maxStep: _indicatorMaxStep,
       child: Transform.scale(
         scale: scale,
-        child: Center(
-          child: widget.camera,
-        ),
+        child: Center(child: widget.camera),
       ),
     );
   }
@@ -150,7 +150,7 @@ class LivenessDetectionStepOverlayWidgetState
 
   Future<void> _handleCompletion() async {
     _updateState();
-    await Future.delayed(const Duration(milliseconds: 500));
+    await Future.delayed(widget.completionDelay);
     widget.onCompleted();
   }
 
@@ -203,9 +203,10 @@ class LivenessDetectionStepOverlayWidgetState
                         Text(
                           'Back',
                           style: TextStyle(
-                              color: widget.isDarkMode
-                                  ? Colors.white
-                                  : Colors.black),
+                            color: widget.isDarkMode
+                                ? Colors.white
+                                : Colors.black,
+                          ),
                         ),
                         Visibility(
                           replacement: const SizedBox.shrink(),
@@ -223,16 +224,19 @@ class LivenessDetectionStepOverlayWidgetState
                         Text(
                           stepCounter,
                           style: TextStyle(
-                              color: widget.isDarkMode
-                                  ? Colors.white
-                                  : Colors.black),
-                        )
+                            color: widget.isDarkMode
+                                ? Colors.white
+                                : Colors.black,
+                          ),
+                        ),
                       ],
                     )
-                  : Text('Back',
+                  : Text(
+                      'Back',
                       style: TextStyle(
-                          color:
-                              widget.isDarkMode ? Colors.white : Colors.black)),
+                        color: widget.isDarkMode ? Colors.white : Colors.black,
+                      ),
+                    ),
             ),
             _buildBody(),
           ],
@@ -294,21 +298,24 @@ class LivenessDetectionStepOverlayWidgetState
                 )
               : ColorFiltered(
                   colorFilter: ColorFilter.mode(
-                      widget.isFaceDetected ? Colors.green : Colors.black,
-                      BlendMode.modulate),
+                    widget.isFaceDetected ? Colors.green : Colors.black,
+                    BlendMode.modulate,
+                  ),
                   child: LottieBuilder.asset(
                     widget.isFaceDetected
                         ? 'packages/flutter_liveness_detection_randomized_plugin/src/core/assets/face-detected.json'
                         : 'packages/flutter_liveness_detection_randomized_plugin/src/core/assets/face-id-anim.json',
                     height: widget.isFaceDetected ? 32 : 22,
                     width: widget.isFaceDetected ? 32 : 22,
-                  )),
+                  ),
+                ),
         ),
         const SizedBox(width: 16),
         Text(
           widget.isFaceDetected ? 'User Face Found' : 'User Face Not Found...',
-          style:
-              TextStyle(color: widget.isDarkMode ? Colors.white : Colors.black),
+          style: TextStyle(
+            color: widget.isDarkMode ? Colors.white : Colors.black,
+          ),
         ),
       ],
     );